Refactor Service
